+++ android_resources_checker/app.py
@@ -1,21 +1,29 @@
+import sys
+
 from rich.console import Console
 from rich.prompt import Confirm
 
+from android_resources_checker.validator import UnusedResourcesException
+
 
 class Application(object):
-    def __init__(self, resources_fetcher, resources_modifier, analyzer, reporter):
+    def __init__(
+        self, resources_fetcher, resources_modifier, analyzer, reporter, validator
+    ):
         self.resources_fetcher = resources_fetcher
         self.resources_modifier = resources_modifier
         self.analyzer = analyzer
         self.reporter = reporter
+        self.validator = validator
 
-    def execute(self, app_path, clients=None):
+    def execute(self, app_path, clients, check):
         self.reporter.apps(app_path, clients)
 
         app_name = app_path.split("/")[-1]
 
         # fetch resources data
         console = Console()
+
         self.reporter.resources_processing_started()
         with console.status("[bold green]Processing project resources..."):
             packaged_resources = self.resources_fetcher.fetch_packaged_resources(
@@ -40,6 +48,15 @@ class Application(object):
         # report
         self.reporter.reporting_started(analysis)
         self.reporter.report(analysis)
+
+        # check
+        if check:
+            try:
+                self.validator.validate(analysis)
+            except UnusedResourcesException as e:
+                paths = "\n".join([r.filepath for r in e.unused_resources])
+                self.reporter.error(f"\nUnused Resources have been found!\n{paths}")
+                sys.exit(1)
 
         # optional resource deletion
         delete_unused_resources = Confirm.ask("Delete unused resources?")